requirements.txt
fastapi
pydantic
redis
numpy
notion-client
sentence-transformers
python-dotenv
uvicorn
prefect
textblob
gradio